+++ README.md
@@ -80,6 +80,10 @@ UI에 사용하는 interface는 `src/models`에 정의한다.
 
 UI에 사용하지 않고 서버와의 데이터 통신에만 사용되는 DTO는 `stores` 내의 스토어 명과 동일한 이름의 타입(`d.ts`)을 정의하여 선언한다
 
+### Services
+
+객체 지향으로 status를 활용하는 함수들의 경우, class 타입의 서비스로 만들어 활용한다.
+
 ## 🚀 빌드
 
 [빌드 프로세스들을 이곳에서 관리합니다. ](docs/BUILD.md)
