+++ src/components/molecules/FeedItemComponent.tsx
@@ -4,7 +4,7 @@ import { useObserver } from 'mobx-react-lite'
 import React, { FC } from 'react'
 import { useStore } from '../../hooks/use-store'
 import { IFeed } from '../../models/feed'
-import { route } from '../../route'
+import { route } from '../../services/route-service'
 import { OverflowMenuIcon } from '../atoms/OverflowMenuIconComponent'
 import { Profile } from '../atoms/ProfileComponent'
 import { TextBase } from '../atoms/TextBaseComponent'
