+++ src/components/molecules/MypageProfileComponent.tsx
@@ -2,7 +2,7 @@ import { IonAvatar, IonButton, useIonViewWillEnter } from '@ionic/react'
 import { useObserver } from 'mobx-react-lite'
 import React from 'react'
 import { useStore } from '../../hooks/use-store'
-import { route } from '../../route'
+import { route } from '../../services/route-service'
 import { Spinner } from '../atoms/SpinnerComponent'
 import { TextLg } from '../atoms/TextLgComponent'
 import { TextXxl } from '../atoms/TextXxlComponent'
