+++ src/components/molecules/TradeItemComponent.tsx
@@ -4,7 +4,7 @@ import { useObserver } from 'mobx-react-lite'
 import React from 'react'
 import { useStore } from '../../hooks/use-store'
 import { ITrade } from '../../models/trade'
-import { route } from '../../route'
+import { route } from '../../services/route-service'
 import { TradeStore } from '../../stores/trade-store'
 import { OverflowMenuIcon } from '../atoms/OverflowMenuIconComponent'
 import { Profile } from '../atoms/ProfileComponent'
