+++ src/components/organisms/SignUpCommunityComponent.tsx
@@ -3,7 +3,7 @@ import { checkmark } from 'ionicons/icons'
 import { useObserver } from 'mobx-react-lite'
 import React, { FC } from 'react'
 import { useStore } from '../../hooks/use-store'
-import { route } from '../../route'
+import { route } from '../../services/route-service'
 import { SubmitButton } from '../atoms/SubmitButtonComponent'
 import { TextLg } from '../atoms/TextLgComponent'
 import { TextXxl } from '../atoms/TextXxlComponent'
