+++ src/components/organisms/SignUpFormComponent.tsx
@@ -3,7 +3,7 @@ import React, { FC, useRef } from 'react'
 import { useForm } from 'react-hook-form'
 import { useStore } from '../../hooks/use-store'
 import { ISignUpForm } from '../../models/sign-up'
-import { route } from '../../route'
+import { route } from '../../services/route-service'
 import { InputNormal } from '../atoms/InputNormalComponent copy'
 import { InputPassword } from '../atoms/InputPasswordComponent'
 import { Spinner } from '../atoms/SpinnerComponent'
